split mockVue afrom mockApp
